refactor: 상수 추출, 컨벤션 적용, 테스트 코드 디렉토리 정리
