+++ src/main/java/lotto/domain/ticket/LottoTicket.java
@@ -8,6 +8,7 @@ import java.util.Set;
 public class LottoTicket {
     public static final int LOTTO_PRICE = 1000;
     private static final String WRONG_SIZE_EXCEPTION_MESSAGE = "로또 번호의 갯수가 %d개로 올바르지 않습니다.";
+
     private Set<LottoBall> lottoBalls;
 
     public LottoTicket(Set<LottoBall> lottoBalls) {
