+++ src/main/java/lotto/view/InputView.java
@@ -5,10 +5,12 @@ import lotto.exception.ConvertFailException;
 import java.util.Scanner;
 
 public class InputView {
-    private final static Scanner scanner = new Scanner(System.in);
+    private static final Scanner scanner = new Scanner(System.in);
     private static final String INPUT_MONEY_MESSAGE = "구입 금액을 입력해 주세요.";
     private static final String INPUT_WINNING_NUMBER_MESSAGE = "지난 주 당첨 번호를 입력해 주세요.";
     private static final String INPUT_BONUS_NUMBER_MESSAGE = "보너스 볼을 입력해 주세요.";
+    private static final String NULL_OR_EMPTY_INPUT_MESSAGE = "입력값이 비어있습니다.";
+    private static final String INVALID_NUMBER_INPUT_MESSAGE = "잘못된 숫자 입력";
 
     public static int inputBettingMoney() {
         System.out.println(INPUT_MONEY_MESSAGE);
@@ -34,7 +36,7 @@ public class InputView {
 
     private static void validate(String input) {
         if (isEmpty(input)) {
-            throw new IllegalArgumentException("입력값이 비어있습니다.");
+            throw new IllegalArgumentException(NULL_OR_EMPTY_INPUT_MESSAGE);
         }
     }
 
@@ -46,7 +48,7 @@ public class InputView {
         try {
             return Integer.parseInt(input);
         } catch (NumberFormatException e) {
-            throw new ConvertFailException("잘못된 숫자 입력");
+            throw new ConvertFailException(INVALID_NUMBER_INPUT_MESSAGE);
         }
     }
 }
