+++ src/test/java/lotto/domain/result/win/prize/PrizeGroupTest.java
@@ -1,7 +1,6 @@
-package lotto.domain;
+package lotto.domain.result.win.prize;
 
 import lotto.domain.result.LottoResult;
-import lotto.domain.result.win.prize.PrizeGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
